Add DB Restore to Maintenance
The function to restore a backup has been added. The "pending" backups are now deleted after the restore.
A backup is also created before a restore, but it is also deleted after the restore. When restoring, the existing database is overwritten by the backup.

CSS adjustments were made to eliminate errors in darkmode and lightmode.