Maintenance
The source of a used GIF added to README.
The source of the used Fonts added to README.
Another used tool was added to the "Powered by" section.
More translation work was done and the technique used to do the manual Nmap-scans was changed to include a loading animation.
A speedtest has been added to the detailed view of the "Internet" device. This is also located under the "Nmap" tab.