difficult to describe
I want to display an assignment of individual network nodes to specific network devices (which device is connected to which switch, router or wlan).

It is still in the experimental phase, but is already usable in principle. Whether this "feature" will make it into the main fork, I can't say yet.

Wegen der aktuellen Entwicklungsphase ist dieser teil ausschließlich deutsch.