Refactor of all functions
Added temporary library "warnings" to the information at the top, this is to facilitate deprecation warnings for guess_type and guess_icon calls until they are fully removed and to remind developers to remove this dependency as soon as it becomes no longer necessary.

Edited/added docstrings for consistency and readability

Edited functions for debug logging, improvements in maintainability, error handling, and improved SQL handling throughout.

Reformatted functions and their arguments for improved readability.

Added handler for if the module is ran directly.